+++ src/__tests__/linter.jest.test.ts
@@ -1,10 +1,15 @@
 import { normalize } from '@stoplight/path';
 import { DiagnosticSeverity } from '@stoplight/types';
 import * as path from 'path';
+import { isOpenApiv3 } from '../formats';
 import { httpAndFileResolver } from '../resolvers/http-and-file';
 import { Spectral } from '../spectral';
 
+import { readRuleset } from '../rulesets';
+import { IRuleset, RulesetExceptionCollection } from '../types/ruleset';
+
 const customFunctionOASRuleset = path.join(__dirname, './__fixtures__/custom-functions-oas-ruleset.json');
+const customOASRuleset = path.join(__dirname, './__fixtures__/custom-oas-ruleset.json');
 const customDirectoryFunctionsRuleset = path.join(__dirname, './__fixtures__/custom-directory-function-ruleset.json');
 
 describe('Linter', () => {
@@ -59,6 +64,22 @@ describe('Linter', () => {
         }),
       ]),
     );
+  });
+
+  it('should support require calls', async () => {
+    await spectral.loadRuleset(customFunctionOASRuleset);
+    expect(
+      await spectral.run({
+        info: {},
+        paths: {},
+      }),
+    ).toEqual([
+      expect.objectContaining({
+        code: 'has-bar-get-operation',
+        message: 'Object does not have undefined property',
+        path: ['paths'],
+      }),
+    ]);
   });
 
   it('should respect the scope of defined functions (ruleset-based)', async () => {
@@ -168,6 +189,225 @@ describe('Linter', () => {
           }),
         ]),
       );
+    });
+  });
+
+  describe('Exceptions handling', () => {
+    it('should ignore specified rules violations in a standalone document', async () => {
+      await spectral.loadRuleset(customOASRuleset);
+      spectral.registerFormat('oas3', isOpenApiv3);
+
+      const res = await spectral.run(
+        {
+          openapi: '3.0.2',
+          info: 17,
+        },
+        {
+          resolve: {
+            documentUri: '/test/file.json',
+          },
+        },
+      );
+
+      expect(res.length).toBeGreaterThan(0);
+
+      expect(res).not.toContainEqual(
+        expect.objectContaining({
+          code: 'info-contact',
+        }),
+      );
+
+      expect(res).not.toContainEqual(
+        expect.objectContaining({
+          code: 'info-description',
+        }),
+      );
+
+      expect(res).not.toContainEqual(
+        expect.objectContaining({
+          code: 'oas3-api-servers',
+        }),
+      );
+    });
+
+    it('should not swallow results', async () => {
+      // Cf. https://github.com/stoplightio/spectral/issues/1018
+
+      const document = {
+        openapi: '3.0.2',
+        paths: {
+          '/a.one': { get: 17 },
+          '/a.two': { get: 18 },
+          '/a.three': { get: 19 },
+          '/b.one': { get: 17 },
+          '/b.two': { get: 18 },
+          '/b.three': { get: 19 },
+        },
+      };
+
+      await spectral.loadRuleset(customOASRuleset);
+      spectral.registerFormat('oas3', isOpenApiv3);
+
+      const res = await spectral.run(document, {
+        resolve: {
+          documentUri: '/test/file.json',
+        },
+      });
+
+      expect(res).toEqual(
+        expect.arrayContaining([
+          expect.objectContaining({
+            code: 'operation-2xx-response',
+            message: 'Operation must have at least one `2xx` response.',
+            path: ['paths', '/a.one', 'get'],
+            source: '/test/file.json',
+          }),
+          expect.objectContaining({
+            code: 'operation-2xx-response',
+            message: 'Operation must have at least one `2xx` response.',
+            path: ['paths', '/a.three', 'get'],
+            source: '/test/file.json',
+          }),
+          expect.objectContaining({
+            code: 'operation-2xx-response',
+            message: 'Operation must have at least one `2xx` response.',
+            path: ['paths', '/b.one', 'get'],
+            source: '/test/file.json',
+          }),
+          expect.objectContaining({
+            code: 'operation-2xx-response',
+            message: 'Operation must have at least one `2xx` response.',
+            path: ['paths', '/b.two', 'get'],
+            source: '/test/file.json',
+          }),
+        ]),
+      );
+
+      expect(res).toEqual(
+        expect.not.arrayContaining([
+          expect.objectContaining({
+            code: 'operation-2xx-response',
+            message: 'Operation must have at least one `2xx` response.',
+            path: ['paths', '/a.two', 'get'],
+            source: '/test/file.json',
+          }),
+        ]),
+      );
+
+      expect(res).toEqual(
+        expect.not.arrayContaining([
+          expect.objectContaining({
+            code: 'operation-2xx-response',
+            message: 'Operation must have at least one `2xx` response.',
+            path: ['paths', '/b.three', 'get'],
+            source: '/test/file.json',
+          }),
+        ]),
+      );
+    });
+
+    describe('resolving', () => {
+      const document = {
+        openapi: '3.0.2',
+        components: {
+          schemas: {
+            TheLocalType: {
+              $ref: './__fixtures__/exceptions.remote.oas3.yaml#/components/schemas/TheRemoteType',
+            },
+          },
+        },
+      };
+
+      let testRuleset: IRuleset;
+
+      beforeAll(async () => {
+        testRuleset = await readRuleset(path.join(__dirname, './__fixtures__/exceptions.resolving.ruleset.json'));
+      });
+
+      const opts = {
+        resolve: {
+          documentUri: path.join(__dirname, './foo.json'),
+        },
+      };
+
+      const extractExceptionFrom = (ruleset: IRuleset, name: string, position: number): RulesetExceptionCollection => {
+        const exceptions = {};
+        const key = Object.keys(ruleset.exceptions)[position];
+        expect(ruleset.exceptions[key]).toEqual([name]);
+        exceptions[key] = ruleset.exceptions[key];
+
+        return exceptions;
+      };
+
+      it('should ignore specified rules violations in a referenced document', async () => {
+        spectral = new Spectral({ resolver: httpAndFileResolver });
+        spectral.registerFormat('oas3', isOpenApiv3);
+
+        const rules = {
+          'strings-maxLength': testRuleset.rules['strings-maxLength'],
+          'oas3-schema': testRuleset.rules['oas3-schema'],
+        };
+
+        spectral.setRuleset({ rules, exceptions: {}, functions: {} });
+
+        const first = await spectral.run(document, opts);
+
+        expect(first).toEqual([
+          expect.objectContaining({
+            code: 'strings-maxLength',
+          }),
+          expect.objectContaining({
+            code: 'oas3-schema',
+          }),
+        ]);
+
+        const exceptions = extractExceptionFrom(testRuleset, 'strings-maxLength', 0);
+
+        spectral.setRuleset({ rules, exceptions, functions: {} });
+
+        const second = await spectral.run(document, opts);
+
+        expect(second).toEqual([
+          expect.objectContaining({
+            code: 'oas3-schema',
+          }),
+        ]);
+      });
+
+      it('should ignore specified rules violations in "resolved=false" mode', async () => {
+        spectral = new Spectral({ resolver: httpAndFileResolver });
+        spectral.registerFormat('oas3', isOpenApiv3);
+
+        const rules = {
+          'no-yaml-remote-reference': testRuleset.rules['no-yaml-remote-reference'],
+          'oas3-schema': testRuleset.rules['oas3-schema'],
+        };
+
+        spectral.setRuleset({ rules, exceptions: {}, functions: {} });
+
+        const first = await spectral.run(document, opts);
+
+        expect(first).toEqual([
+          expect.objectContaining({
+            code: 'oas3-schema',
+          }),
+          expect.objectContaining({
+            code: 'no-yaml-remote-reference',
+          }),
+        ]);
+
+        const exceptions = extractExceptionFrom(testRuleset, 'no-yaml-remote-reference', 1);
+
+        spectral.setRuleset({ rules, exceptions, functions: {} });
+
+        const second = await spectral.run(document, opts);
+
+        expect(second).toEqual([
+          expect.objectContaining({
+            code: 'oas3-schema',
+          }),
+        ]);
+      });
     });
   });
 });