+++ src/rulesets/mergers/__tests__/functions.jest.test.ts
@@ -1,16 +1,10 @@
 import { RuleCollection } from '../../../types';
 import { RulesetFunctionCollection } from '../../../types/ruleset';
 import { mergeFunctions } from '../functions';
-const nanoid = require('nanoid/non-secure');
 
 jest.mock('nanoid/non-secure');
 
 describe('Ruleset functions merging', () => {
-  beforeEach(() => {
-    let seed = 0;
-    (nanoid as jest.Mock).mockImplementation(() => `random-id-${seed++}`);
-  });
-
   it('re-writes function names', () => {
     const target = {};
     const sources: RulesetFunctionCollection = {
@@ -18,6 +12,7 @@ describe('Ruleset functions merging', () => {
         name: 'foo',
         code: 'foo()',
         schema: null,
+        source: null,
       },
     };
 
@@ -37,6 +32,7 @@ describe('Ruleset functions merging', () => {
       name: 'foo',
       code: 'foo()',
       schema: null,
+      source: null,
     });
   });
 
@@ -46,6 +42,7 @@ describe('Ruleset functions merging', () => {
         name: 'foo',
         code: 'foo()',
         schema: null,
+        source: null,
       },
     };
     const sources: RulesetFunctionCollection = {
@@ -53,6 +50,7 @@ describe('Ruleset functions merging', () => {
         name: 'foo.c',
         code: 'foo.a()',
         schema: null,
+        source: 'foo',
       },
     };
 
@@ -62,11 +60,13 @@ describe('Ruleset functions merging', () => {
       name: 'foo.c',
       code: 'foo.a()',
       schema: null,
+      source: 'foo',
     });
     expect(target).toHaveProperty('foo', {
       name: 'foo.c',
       ref: 'random-id-0',
       schema: null,
+      source: 'foo',
     });
   });
 
@@ -76,6 +76,7 @@ describe('Ruleset functions merging', () => {
         name: 'foo',
         code: 'foo()',
         schema: null,
+        source: null,
       },
     };
 
@@ -84,11 +85,13 @@ describe('Ruleset functions merging', () => {
         name: 'foo',
         code: 'a.foo.c();',
         schema: null,
+        source: null,
       },
       bar: {
         name: 'bar',
         code: 'bar()',
         schema: null,
+        source: null,
       },
     };
 
@@ -112,11 +115,13 @@ describe('Ruleset functions merging', () => {
       name: 'foo',
       code: 'a.foo.c();',
       schema: null,
+      source: null,
     });
     expect(target).toHaveProperty('random-id-1', {
       name: 'bar',
       code: 'bar()',
       schema: null,
+      source: null,
     });
   });