+++ src/utils/__tests__/sortResults.test.ts
@@ -1,6 +1,6 @@
-import { DiagnosticSeverity } from '@stoplight/types';
+import { DiagnosticSeverity, IPosition } from '@stoplight/types';
 import { IRuleResult } from '../../types';
-import { compareResults, sortResults } from '../sortResults';
+import { comparePosition, compareResults, sortResults } from '../sortResults';
 
 const results: IRuleResult[] = [
   {
@@ -203,4 +203,32 @@ describe('compareResults', () => {
       expect(compareResults({ ...input, code: tc.one }, { ...input, code: tc.another })).toEqual(tc.expected);
     });
   });
+});
+
+const buildPosition = (line: number, char: number): IPosition => {
+  return { line, character: char };
+};
+
+describe('comparePosition', () => {
+  const positionTestCases = [
+    [2, 2, 1, 1, 1],
+    [2, 1, 1, 1, 1],
+    [1, 2, 1, 1, 1],
+
+    [1, 1, 1, 1, 0],
+    [1, 2, 1, 2, 0],
+    [2, 1, 2, 1, 0],
+    [2, 2, 2, 2, 0],
+
+    [1, 1, 1, 2, -1],
+    [1, 1, 2, 1, -1],
+    [1, 1, 2, 2, -1],
+  ];
+
+  test.each(positionTestCases)(
+    'should properly order locations (%i, %i) vs (%i, %i)',
+    (leftLine, leftChar, rightLine, rightChar, expected) => {
+      expect(comparePosition(buildPosition(leftLine, leftChar), buildPosition(rightLine, rightChar))).toEqual(expected);
+    },
+  );
 });